+++ docs/src/App.js
@@ -55,7 +55,6 @@ export default function App() {
 						<Address
 							asJSON={true}
 							value={ data }
-							preferredCountries={ ['US', 'CA', 'JP'] }
 							onChange={ value => setData(value) }
 							classNames={ {
 								select: "block w-full mt-1 rounded-md border-gray-300 shadow-sm focus:border-indigo-300 focus:ring focus:ring-indigo-200 focus:ring-opacity-50",