+++ CHANGELOG.md
@@ -7,6 +7,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ## [Unreleased]
 
 ### Added
+
+- Added support to auto-detect country based on local timezone
+
 ### Changed
 ### Deprecated
 ### Removed